Fix BT memory leak. Initiate cleanup

+++ src/node1/slider.h
@@ -1,11 +1,14 @@
 #ifndef __SLIDER_H__
 #define __SLIDER_H__
 
+#include "adc.h"
+
 typedef struct
 {
-    int8_t left_pos;
-    int8_t right_pos;
+    int16_t left_pos;
+    int16_t right_pos;
 } slider_pos_t;
 
 slider_pos_t slider_read_pos(void);
+
 #endif
